+++ deltafi-ui/CHANGELOG.md
@@ -4,12 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.99.4]
 
 ### Changed
 
 - Added delete button to System Snapshots page (#295)
 - Added Cancel menu option to DeltaFile Viewer (#293)
+- Added Requeue Count minimum filter to DeltaFile Search Page (#294)
+- Added Egress Flow filter to DeltaFile Search Page (#302)
 
 ## [0.99.3]
 
@@ -471,7 +473,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - NiFi link from Dashboard
 - Requested resource information from System Metrics
 
-[unreleased]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.99.3...main
+[unreleased]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.99.4...main
+[0.99.4]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.99.3...0.99.4
 [0.99.3]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.99.2...0.99.3
 [0.99.2]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.99.1...0.99.2
 [0.99.1]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.99.0...0.99.1